Mostly configurable fshow
Introduces a string list in the config.yml for /f show output
customization.

Lang file for textual changes
Config file to determine if we’re going to send the specific part of
the f show output.

The show command could use some optimization! If we’re not sending
enemies, then why build the entire message first…?

Will work on those optimizations later, this is just initial
implementation of concept.