Variable overhaul
Introduces a new system of handling variables from config that get
replaced with in-game values. Previously we had 3 different sections of
code handling the same thing and each did not do a particularly great
job (missing variables) not to mention whenever we wanted to add a new
variable we had to do so in 3 different places.

Additionally, revamped the config and renamed several existing variables
to fit a specific style (words seperated by hyphen, all lowercase, and a
descriptive name). Went through all of the config comments and updated
them. Added a giant list of usable variables to the bottom of config.