Fix faction DTR not starting from 0
Since lastDtrUpdateTime was never initialized, updateDTR() was using a
huge delta which shot the DTR to a huge number, which was then limited
by our built in detection.

This had the effect of DTR starting at 0, but then when the faction gets
it's first updateDTR() call, DTR went to the max.

Fixes https://github.com/drtshock/HCFProblems/issues/123