Adds option for minimalistic /f show
Re adds the option for a minimal f show, which is false by default.

Any balance, land value, faction home, land refund, and all fancy
variables will be have the line they are present on ignored if they are
"empty" and minimal show is enabled. Empty is if for example, 0 players
are online; we wont bother to add a line in there saying Online(0/x):

Resolves https://github.com/drtshock/HCFProblems/issues/160