Adds {faction-max} variable
Adds a simple way to get the max players a faction can have. Previously
you'd have to manually match the values in config.yml and conf.json if
you wanted to display to the end user in /f show what the max was. This
reduces the possibility of errors somewhere in that process.

https://github.com/drtshock/HCFProblems/issues/276